refactor object detection transform
